fixup! Fixed bug when spikes vector is empty  - std:min/max_element() dereference end iterator Added unsuported functions from C API

+++ src/reports/library/implementation_interface.hpp
@@ -85,8 +85,8 @@ namespace detail {
         };
 
         inline static void sort_spikes(std::vector<double>& spikevec_time, std::vector<int>& spikevec_gid) {
-            double lmin_time = std::numeric_limits<int>::max();
-            double lmax_time = std::numeric_limits<int>::min();
+            double lmin_time = std::numeric_limits<double>::max();
+            double lmax_time = std::numeric_limits<double>::min();
             if(!spikevec_time.empty()) {
                 lmin_time = *(std::min_element(spikevec_time.begin(), spikevec_time.end()));
                 lmax_time = *(std::max_element(spikevec_time.begin(), spikevec_time.end()));
